+++ DS Project/Rest/Order.h
@@ -12,7 +12,7 @@ protected:
 	int ID;                             //Each order has a unique ID (from 1 --> 999 )
 	ORD_TYPE type;	                	//order type: Normal, Vegan, VIP
 	ORD_STATUS status;	                //waiting, in-service, done
-	int Dishes;	                        //The number of dishes in that order            
+	double Dishes;	                        //The number of dishes in that order            
 	double totalMoney;	                //Total order money
 	double Price;                       //the total price for one dish
 	int ArrTime, ServTime, FinishTime;	//arrival, service start, and finish times
